Corrección de errores en los tipos de datos

+++ src/main/java/co/edu/javeriana/jpa_example2/model/Partida.java
@@ -17,7 +17,7 @@ public class Partida {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int tiempoLimite;
+    private Long tiempoLimite;
     private double gananciaMinima;
     private LocalDateTime tiempoInicio;
     private Long tiempoActual;
